some cleaning
